+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- A static Issue Launch.NOT_ISSUE to use in agents and avoid issue creation duplication
+- StatusEvaluation class with an `evaluateStatus` method
 
 ## [5.0.17]
 ### Fixed